Swapped out Generic Value
Now it's using DSTs. It's not as neat, but it'll be easier to generate
code this way. This way turns out to be more flexible in general,
and I may need that. I can always get creative and go back to the
generics if I want to.

I developed a better way of initializing the proxy types too.