Super Magic References
Got assignment working -- I think. It works for simple locals  anyway,
not sure how it's going to pan out for local structs.

References are a thing now. At least on the command line. It doesn't
look like a reference though. I don't know if I'll need to address that
or not.