Lot's o' good stuff ☘️
So, firstly the tests are all now passing! Yay!

To accomplish that I shuffled some code around and compared the error
output of dwarf to the contents of a `test.stderr` file.

I also needed to get the or operater `||` working, which was pretty
simple really.

Finally, there's a test and code coverage script attached to this, so expect
fun stuff there.