Mostly Error Handling
I created a ChaChaErrorReporter (like in Dwarf) and am using it to
print errors for lambda. I probably need to do the same thing for
the dwarf binary. It's also looking like I'm going to be adding span
information to a lot of errors in chacha.

I fixed the exrpession counting reporting too.