Working on the Book 📔
I started working on getting the code in the book working. (I need to
look into hacking the mdbook test command to test dwarf.) This led to
one thing and another and next thing you know you've been hacking on
lot's of stuff.

This seemed like a decent enough place to stop.