... created a bug, fixed two bugs.
My last fix was probably just me undoing the previous "fix".
Now everything is really fixed.

There's some code that gets called in the extruder to inter a method call.
Like all expression interring functions, this one needs to return an
expression and a type. The issue is the type. I never went to the trouble
to figure out what the return type was. I either punted and said unknown,
which helped things if I allow that to typecheck. Or, I again punted and
said the type was the type upon which the method call was being made.
This helped with method chaining..

The right thing to do is to look that stuff up. If it's a UDT I have all
that info in LuDog. If it's a built-in like a string, then I just do a
match on the method name and return the right type. Seems to be working.

I turned off allowing unknown types to typecheck. Still works.

This all has the nice side effect of catching bogus calls on strings
early. This should be pretty easily extended to the chacha stuff, but
tbh, I'm not sure that it belongs in there. Chacha is an extern. So
really we need to write a definition for it and import that with a use
item.

So much work to do...