Got λ's?
Sweet closures, sweet closures. 🎵

So yeah, I've got a weak version of a closure. I have not tested extensively,
and assume that they are fundamentally broken. That said, they parse, there
is a place for them in the model, and they execute.

More tests to come. Just needed to commit before things got out of hand.