Fixed Harness Error Diff
I broke it last commit. This fixes it, and makes it better by refactoring
that diff code into a function.

There's also a modification to the stdout of an impl block test. That's
actually going to fail, because it fixes a test for something that
changed in the next commit.