Better Struct Expressions
I was adding attributes to structs and I realized that there was yukky
with struct expressions. We were checking for the existence of fields,
and typechecking them during runtime.

I moved all of this to extrusion time, and made the error message much
better. The old one was in fact pointing at the wrong thing.

I also seem to have missed a file in a recent commit. 😞