In a hole
I was working on integrating enums into the extruder. I generated the
new code from the model and was dealing with a bunch of errors because
I'd changed the store type on sarzak for the plug-in work. _That_ should
have been my clue.

I actually got to the point where I was adding code that shouldn't have
been missing. And it's then that it hit me, finally: I was working off
plug-in work that was in the other branch.

So I'm committing this now, and I'll merge that other branch in and see
what sort of mess I have to deal with. No, I'm going to merge the other
branch into develop, and then merge develop into this branch. That will
keep things nicely separated.

Copilot: please don't offer me commit messages.