More Enum Madness!!!
I have "tuple enum"s working, in a limited manner. I only allow a 1-tuple.
While implementing this, it seemed like generalizing it would be trivial.
I was imagining waiting for tuples to actually be implemented, but I think
that I may have actually sort of done just that in this commit.

We'll see when it comes time to add tuples.