Models and more
I'm committing this work from yesterday before I get started seriously
this morning.

Yesterday I managed to get the model files loading from the plugin, rather
than from the command line. I was quite pleased with this.

I was trying to get everything working end-to-end in the REPL when I ran
into problems. The problems are mutex woes, and they are because I am
trying to inter a function into the store and then run it.

Backing up, in the REPL when I `use` a plugin module it works just fine.
The problem is that the newly interred structs and whatnot are not in
the interpreter's memory. So when you invoke a static method on any of
the types the lookup failed. I moved the lookup to use LuDog, inter the
function and execute it, and I hit the mutex wall.

I think anyway. I think because it only happens in the REPL. I'm currently
in the midst of passing a vec of dirty types that need to be interred
before the interpreter evaluates anything.