This is a big one
I removed the type from the return value for most of the interpreter
functions. It was unnecessary, and got annoying.

I also fixed a problem with the mutexes, introduced by the plugin work.
Sadly none of my tests at the time caught it. I wrote another test thinking
I was going to uncover a bug I thought I found. Instead this. Good deal.