For loop iter-thing is now in block
The lhs value for the for loop iterator is now properly put into the block
and now typechecking can proceed as per necessary. This fixed a problem
where the loop variable was not being recognized as a valid struct.

While I was in there I also updated the color/printing of types on the
dwarf side of the wall. I'd really like to unify those.

I had to play with index and range some. Those need looking at.