Started in on async events to instances
Found a bug where if you have a type that's the same name as a type in
sarzak, you will bomb. Fixed it.

Also messed about with finding double span assignments, which are actually
ok -- the last span is probably the one we want. But then again, I don't
like it much.