Sigh, too much again.
There's a bunch of stuff in here that really needs to get merged into
develop. Lots of minor things that I've already forgotten about.

Major is that struct generics are looking really good. Enums are in
decent shape too. They both need a lot more work though.

I've also got a good start on futures. I was on my way to returning a
Future type from async blocks and functions -- just like Rust, yes.
I was then going to key off of that to implement an 8.and_then()`
method to get at the results of the future.