Damn.
This is a real melange. There is a bunch of stuff related to async,
threads, and timers. But I sort of got frustrated, and I needed a
break.

So I started working on a plugin to do http requests. That led to some
changes is the xtask thing that I use to build the plugins. Misha says
I need to commit more often.