Too tired to think
Not really sure all that I did, but it's running sort of. There are
panics in the threads because the queue is going away before they are
done waiting on it.

The thing is blocking on threads during sleep. I think I need to spawn
the timers on the main executor. I seem to remember something about
that. Even with enough threads, it's about twice as slow as I expect it
to be. Well, at 42 executors. It scales with the count.