Async (#13)
* Started in on async events to instances

Found a bug where if you have a type that's the same name as a type in
sarzak, you will bomb. Fixed it.

Also messed about with finding double span assignments, which are actually
ok -- the last span is probably the one we want. But then again, I don't
like it much.

Also, we are now parsing `async fn`.

* Working an async and got sidetracked

Ended up fixing some bugs, and making some stuff I needed work. Mandelbrot
is still busted I'm afraid.

* Getting enums working better

I have an async program for testing and the enums were too broken to
work. So I fixed them.

* remove dbg statements

* cleanup

* Parse await

That's about all. Need to merge in stashed async changes to integrate
latest sarzak model changes.

* Merge stashed changes.

* Sigh, too much again.

There's a bunch of stuff in here that really needs to get merged into
develop. Lots of minor things that I've already forgotten about.

Major is that struct generics are looking really good. Enums are in
decent shape too. They both need a lot more work though.

I've also got a good start on futures. I was on my way to returning a
Future type from async blocks and functions -- just like Rust, yes.
I was then going to key off of that to implement an 8.and_then()`
method to get at the results of the future.

* Lots 'o files

Most of these are test failure output that changed when I hacked on
Ariadne. I need to submit a PR for that...

The highlight is that I have a broken, half-assed async implemented.

* Backed out broken-ness

The future thing I had broke tests. It was bogus anyway.

* Fix async block bug

* Async Working

As far is it's been tested anyway.

* Tests and Feature Flags

Got async working in the test harness.
Got async behind feature flags.

* Spawn?

Added spawn, although it's wrong, as is async. Some other misc
junk/clean-up.

* Let there be Parallelism!

Problem is that everything is running in parallel -- I need to get await
working while allowing other things to run.

* Need to put this away before I do the next stupid thing.

* Got threads running.

Something not quite right though. Missing some print statements.

* Thread is now finishing it's work before going away.

* Damn.

This is a real melange. There is a bunch of stuff related to async,
threads, and timers. But I sort of got frustrated, and I needed a
break.

So I started working on a plugin to do http requests. That led to some
changes is the xtask thing that I use to build the plugins. Misha says
I need to commit more often.

* It's Tuesday

I now have dwarf/extruder errors carrying around the file from whence
they came.

I got pretty far hacking in an http client plugin. I'm going to abandon
that effort in favor of just hanging it off of chacha for now. I'll
come back to the plug-in thing and do it right later.

* Not sure really.

Bacon wasn't testing and there were errors, and now they are fixed. I
don't know how old they were though. So it could be that the changes
in this commit are what broke the tests.

I think that I was working on hacking in an http client.

* Fuck me.

I have no idea what all I did. However, the good news is that my super-
awesome-delayed-task-starting code is working now! That's one of the
sweetest pieces of code I've ever written.

* I have no clue

I'm stuck with segfault that I don't know how to fix.

* Changed dwarf to accept a thread count argument.

Most likely more too. :(

* Put the threads back in

* Yay -- timely commit, shutdown and generics

I figured out a slick way to make the threads all shutdown when I want
them to, without blocking or spinning.

I also made the executor and it's task generic -- gonna make a crate.

* My shutdown code

It was fine before I committed, I must have changed it back to what
didn't work, and then committed it. IT's right now.

Also messing with the thread count bug.

* Next step for crate

Moved threads out of interpreter into async.

* Too tired to think

Not really sure all that I did, but it's running sort of. There are
panics in the threads because the queue is going away before they are
done waiting on it.

The thing is blocking on threads during sleep. I think I need to spawn
the timers on the main executor. I seem to remember something about
that. Even with enough threads, it's about twice as slow as I expect it
to be. Well, at 42 executors. It scales with the count.

* Motherfucker is working I think

Title says it all.

* Minor junk.

* Misc

Not sure exactly, but for sure this includes moving to using my
uberfoo-async crate.

* moved to failing until I get other problems sorted

* Deadlock Gone

Got rid of the deadlock. Fixed last test. There is still an issue with
the list type. Deal with that next.

* Added MAP to range methods in extruder

* make tracy optional

* Last remaning vestiges of Tracy are now gone.

* Clippy, and makeing async optional.

* Made async optional. Clippy lints.

* Remove Trace, Fix Async Bug

Tracy isn't removed so much as behind a feature.

The bug was that I wasn't blocking on the parent process at the await
point.

* Working on await

I fixed a bug where Future<unknown> printed as Future<<unknown>>.

* Future<unknown> Spawn Problem

I wasn't getting the type of the argument to spawn, and it sucked.

* More Good Stuff

- Gave Value::Vector a type.
- fixed calling an async function
- changed exec lambda to take values instead of arguments
- implemented map on the list type

* Function Signature Parsing

Now processing funcs for their signatures so that we can look up the
type of function when interring a call to it.

* minor changes

* update puteketeke version