Working on Generic Functions
Got pretty far with this. We are passing generics from the parser to
the extruder where functions are using them. Actually calling the function
does not work though.

`let x = id(42)`, where `id` is a generic function, does not work. The
extruder errs with expected `<T>` found `int`. When I try this,
`let x = id::<int>(42)`, the interpreter panics because it's trying to
call a static function or something.

After this of course would be to do a little type inference and make the
bare call work.