Feature/aoc 2023 (#18)
* Std Working

Mostly this was about getting the plugins to work.

* Fix `chacha::parse` and `chacha::eval`

There were no return type definitions for the functions and so the
interpreter was freaking.

* Working on Generic Functions

Got pretty far with this. We are passing generics from the parser to
the extruder where functions are using them. Actually calling the function
does not work though.

`let x = id(42)`, where `id` is a generic function, does not work. The
extruder errs with expected `<T>` found `int`. When I try this,
`let x = id::<int>(42)`, the interpreter panics because it's trying to
call a static function or something.

After this of course would be to do a little type inference and make the
bare call work.

* Advent of Code 2023.1

This is a solution to Advent of Code 2023 problem 1, written in dwarf.

It most importantly contains enhancements to the language needed to
enable the program.

* Advent of Code 2023.2

This is a solution to Advent of Code 2023 Day 2, written in dwarf.

It most importantly contains enhancements to the language needed to
enable the program to be written.

* Got Generic Funcs Working

Some tests too. I basically updated typechecking to pass through anything
generic. Not sure if that's really the right way to go or not.

* Bad bug in integer method call

We weren't actually testing in the match, but just falling through.
Updated aoc tests to have assert statements.

* Added some missing plugin files

* Unicode identifiers

Working on documentation I discovered to my dismay that I was unable to
use an `α` as an identifier. Fixed that shit.

* Update

* Clippy

* Printing Vector's

The output of printing a dwarf list used to be ugly. Now it is not.

* Better string display

I made it so that a string will display with quotes around it when it
is printed.

* Update CI to copy std files to dwarf home.