Got Enum Implementations Working
There's weirdness with enums. We store them without the generic type,
but that doesn't work when we look at the impl, because it could have
any generic parameters at all.

Then there is some weirdness with the type lookup for the impl.
This won't work
    `let e = Option::<string>::None;`
and this will.
    `let e: Option<string> = Option::None;`

Those should be the same type, and they aren't. The former is `Option<string>`
and the latter is `Option`.