So, I did two things that are unrelated...
First off, the VM/compiler stuff. Yup. It's that time. Basically this
was all structural business. I do have a few tests already for the
compiler, which is sweet.

The other thing I did was fix the way an async function call is handled
if async isn't turned on. I created a new error, and all that business.
And it returns that error.

I also did some general cleaning. Mostly wrt async stuff.

Added a test that won't fail because we don't test both run times. I'll
have to get to that later.