Gave VM it's own error type.
This moved VmPanic out of ChaChaError.
I also made a local BubbaError for the VM. I'll do the same for the
compiler.

The end goal is to move the bubba module to a new crate. The rationale
behind this is so that I can make a VM that runs in WASM. Giving it a
little thought though, I do not know if this is even doable.

I think I can manage to get the VM to run in WASM, but I do not know if
I can get the compiler running, since it needs the entire front end.
The parser is probably not a problem, it's the extruder I'm worried
about.

And actually that is not the end of the world.