For Loops -- Sort of
So yeah, I sorted out how to get for loops working in the compiler.
They are only over ranges, which is a huge limitation of course.
I also have no idea how to generalize this over lists and strings.
Of course, writing this has given me some clarity.

I was stressing out over how to generate assembly for general iterators,
but that's not really required. I just need to consider lists and strings.
So I should be able to check the type of the iterating expression and
then generate specialized assembly. That seems like a good starting
point anyway.

This did require a change to the model, bee tee dubs.