Most Excellent Commit
I got else if working in dwarf in general. It turned out to be fairly
trivial. And I was putting it off dreading it. I added it because of a
fib test I wrote in the compiler. Actually, copilot wrote it, and it
contained an else if, so I figured WTF?

Of course the fib test results in a SO. So, there's that.