Plug-in Hacking
This morning I was daydreaming about a demo or something. Was wondering
if it would be possible to create a state, and then fill it with dwarf,
from the REPL. Anyway, I wound up fixing a bug. But there's still something
wrong.

When I create a new instance of a WoogStruct, it's type is Struct,
wrapped in a sarzak Ty. Not correct.