Working on import and module expressions
I've got everything jacked up now for this. The signatures of all the
extruder things are now updated to carry around a list of string/lu_dog
tuples.

Contexts are updated to have a hashmap of these downstream.

Not all is super-great though. There are still places where I'm not looking
up the types. Better to save this off now though.