Imports looking better
I got rid of the idea of multiple lu-dogs since the expression outputs
would not match if the expression was in another db. Instead I updated
the model to have a path for enums and structs.

I still need to clean up the context_stack business. I may need to leverage
that parameter for something that indicates that a type is in scope.