Odds and Ends
I started out wanting to get formatted strings working. That is a can of
worms. I need to rewrite part of parser. Basically I need to parse expressions
in the strings.

So then I tried to get mandelbrot working. I fixed a few bugs in function
calls, but there's still something fishy going on.

I'm going to hack on the VM to remove the recursion from the run loop.
I'll be managing the stack pointer then, and I can be sloppy in the compiler
about cleaning up after myself. This'll probably fix the mandelbrot bug.

I feel sort of guilty doing it this way, but every time I try to sort out
what's going wrong I get sleepy. I'll come back to it.