Plug-in Fun
I'm working toward adding plugins to the VM. I wanted to sanity check
that sarzak and lu_dog were working, as getting http working for real
is my end goal.

Turns out that dwarfc couldn't generate a lu-dog that worked with the
generated plug-in code. So that's what this is about.