Generics are (mostly?) Sorted
I'm not even sure what started me down this path. I guess it's the plugin
stuff. On that road I ran into terrible issues with generics and matching,
and all that fun stuff.

This commit contains not only a working http client plugin, which relies
on very little glue, it also has generics figured out once and for all.