Damn -- that was intense.
This was focused on plugins and imports -- they go hand in hand.

Imports rock now. I've got scopes tracking what's available. I'm storing
everything in LuDog with it's name being the full path so that there's
no ambiguity.

The plugin stuff is pretty seamless as well.

There was a lot of work that went into this, and I apparently have
little to say about it.