Somehow I got ahead of myself
There's a complete mishmash of stuff going on in here.

Most of it is around trying to get plugins working in the VM. The reason
for this commit at this time is that there were some problems getting
the async_http client running under dwarf. I don't know how the test
didn't fail.