VM Optimization
I noticed that a lot of time was spent creating and destroying smart pointers
to values on the stack. So I figured I'd store an enum that was either
a reference or a value. This actually takes more time than just creating
and destroying them in single-threaded mode. However, in multi-threaded
mode, it's twice as fast.

To be fair this is only testing on fib, but that's a good stressor of
function calls.

This can be expanded upon by creating two different push instructions.
As it stands the compiler is always pushing references, and there are
likely as not a lot of places that can be elided.