Async Finally Working
The problem was that values from the program were
being shared across the tasks, and all the locking
and unlocking killed performance, and setup an inter-
thread dependency that was keeping things from running
in parallel.

The specific instruction causing issues was PUSH, which
used to take a RefType. I did this to keep the instructions
small. The size jumped from 16 or 24 to 88. Surprisingly
the performance on benchmarks improved between 14 and 16
percent. These benches are all sync, so that's a nice
side effect. I will remove all RefTypes from the instructions
at some point.