Feature/plugin async paper (#29)
* Removed unused struct.

* Fix Printing in the VM

Clippy had me remove an unnecessary `format!``, and it was wrong.

* Added Method Calls to Format String

Yup. Now you can do ```print(`foo.len()`)``` and it works.

* Tests for Method Call in Format String

See above.

* Fix Lambda Empty Return Type

Returning empty from a lambda wasn't properly pushing the () to the stack
before returning.

Also "fixed" a place an unwrap was causing trouble. It's to do with cargo
test and once_cell.

* Added Async Call Semantics to VM

Everything was a spawn before. Now there is also a call.

* fix test running problem

* Yeesh

- Added a custom value type to the VM. Improved benchmarks from 25%
(fib) to 36% (loop).
- Removed VM from interpreter.

* Async Finally Working

The problem was that values from the program were
being shared across the tasks, and all the locking
and unlocking killed performance, and setup an inter-
thread dependency that was keeping things from running
in parallel.

The specific instruction causing issues was PUSH, which
used to take a RefType. I did this to keep the instructions
small. The size jumped from 16 or 24 to 88. Surprisingly
the performance on benchmarks improved between 14 and 16
percent. These benches are all sync, so that's a nice
side effect. I will remove all RefTypes from the instructions
at some point.

* Comments

* Updated CHANGELOG and a flaky test.

* Cleanup

* update ffi path

* Welcome char!

- Added the char type and char literals to the parser, model, etc.
- Comment cleanup and code refactoring.
- Added tests, and fixed code where necessary.

* Plugin Cleanup

Removed unused `close()` method from `PlugIn` trait.

* Persisting Compiled Source

- We store the compiled source in $DWARF_HOME/compiled.
- The compiled files have an extension of .gp for "Gold Pieces".
- The compiled file is named {hash(path)_file_name}.
- We check timestamps of the source and compiled file to determine if
we need to recompile.
- We are persisting as JSON 🤮 because deserializing bincode panics.

* For Loop Hack

I hacked in a cx register to act as the for loop counter.
It seems to have had substantial effects.

* Mostly Instruction Cleanup

I removed most of the pointers wrapping Strings in the instruction
set.
Cleaned up mandelbrot some, although it's still broken.

* Added install xtask

It installs:
- dwarf
- http plugin
- std lib files

* Changed xtask command to build plug-ins

Changed from plugins to plugin.

* Goofing off

- I added a harness and tests for the VM. These are copied from the interpreter
tests.
- I added an xtask install target.
- I added pretty error messages for the VM
- I added pretty error messages for the compiler

That's probably all.

* A crap-ton of stuff

I touched almost everything getting these tests to run, and sadly there
are for too many that I just decided to punt on.

* Update README, bump version.

* Hash

Mostly this is about changing Value::Struct to directly store a Struct,
and not a RefType<Struct>.
There's some other stuff. I think I fixed how gp files are hashed, as
well as where they are written.
Probably other stuff too.

* Managed to finally fix touching the files

Fixed main to check the gp version and time against the compiled file
and overwrite if newer.

* Just Whatever

I've been hunting down a bug with generics, and I've fixed a lot of stuff
that was broken, but didn't get me where I wanted to be. It's time to
just commit this and move on.

* Update README.md