Working on plugin lambdas
This is super close to working. The issue is that I cannot
access the lambda map from the plugin code. The plugin has
it's own versions of those symbols. I can't seem to get rid
of them.

The next thing to try is moving stuff into a separate crate
that is only included by the main binary code, and not the
plugin code.