Lambdas in plugins!
So yeah, I got lambdas working across the FFI boundary.

I'm storing a table of LambdaPointers and passing around a
handle. I compile a trampoline function to call when the
lambda is called. The calling is done via a channel. The
return value is passed via a different, ephemeral channel.