Remote Lambdas (#32)
* Minor tweaks. Get symbols back.

* Working on plugin lambdas

This is super close to working. The issue is that I cannot
access the lambda map from the plugin code. The plugin has
it's own versions of those symbols. I can't seem to get rid
of them.

The next thing to try is moving stuff into a separate crate
that is only included by the main binary code, and not the
plugin code.

* Refactor Value

* Lambdas in plugins!

So yeah, I got lambdas working across the FFI boundary.

I'm storing a table of LambdaPointers and passing around a
handle. I compile a trampoline function to call when the
lambda is called. The calling is done via a channel. The
return value is passed via a different, ephemeral channel.

* Good things come to those who hack.

- Switched over to hyper. Got static hello world working.
- fixed if let bug, which was related to StoreLocal instr. Turns out
that I needed to make a new instruction for let and match statements.
- part way through making an any list for use with invoke_func.

* Update for AnyList

* Any Type Completion

Got it in there and working.

* One step closer to a server

Some cleanup. Got Result type -> FfiValue working.

* Got HTTP Server Working!

Not much more to say. Lots of changes in http plugin code.
In the dwarf side, I updated the lambda pointer creation to
share values across lambdas. Needed to update stack with
pointers.

* Oopsie -- missed a test. Always run bacon...

* Woot! Lambdas in plugin get to call back in for params

Sounds weird. Stuff in the plugin can't escape really. To call
functions on things that come back requires basically an RPC.
Got that working from code the starts in the plugin.

* minor tweaks.

* Refactoring and VM Trace

Cleaned up the VM a touch. There was code duplication in the
async call and async spawn functions.

I also made trace part of the VM struct, and removed our
dependence on RUST_LOG=vm=trace. This means a new flag on
the binary.

* Refactoring and VM Trace

Cleaned up the VM a touch. There was code duplication in the
async call and async spawn functions.

I also made trace part of the VM struct, and removed our
dependence on RUST_LOG=vm=trace. This means a new flag on
the binary.

* Refactoring and VM Trace

Cleaned up the VM a touch. There was code duplication in the
async call and async spawn functions.

I also made trace part of the VM struct, and removed our
dependence on RUST_LOG=vm=trace. This means a new flag on
the binary.

* Bunch 'o Tests

Trying to get code coverage up so that I can resolve this PR.