Feature/db (#35)
* Debug tweak

* A few things

Added std::fs::File and moved std to the plug-in directory.

Added escapes to strings.

* Remember Two Things

- -c flag to change directory is there, and busted
- updated README pleading for help

* Remember Two Things

- -c flag to change directory is there, and busted
- updated README pleading for help

* Markdown Plugin

Nuff said.

* File stuff working, some clean-up

* Hopefully Close this PR

Some tests for added functionality and fix install.

* Fix CI install step

* Fix CI install step

* Cleanup

Both literal and runtime. The literal was just chasing clippy. The
runtime is that we are no longer interring and thus compiling all
the things that are included in a library. IOW, we are only compiling
the functions that we need.

* HTML in Markdown

Hacked the markdown plugin to allow HTML in markdown.

* Cleanup

Got rid of some dbg! invocations.

Added dotenvy, and that's really for the sql stuff, but whatever.

* A few things

moved plug-ins to plugins.

added sqlx plugin.

* DB Query Working

Got it so that I can map a query onto a type.

* Moving stuff around

Moved some things in the match tests to see if the order caused problems.
Chasing a bug in the db connection match code that breaks rather than
return an error.

* cruft

* Added test for db query.

* Update CI to use postgres.

* Update postgres password.

* Redid SQL plugin API

I made it look more like sqlx syntax where there's
a query, map and fetch_all. I think it's better this
way, but more importantly it proves that I can do it,
and I need this sort of behavior for `bind`.

* W00t!!

Got bindings working on database queries, so I can read and write
now.

* Added env::var env::set_var

These were added to the std plugin.

* Got plot of counter working

Got REST endpoint /counter_data working to serve up the last 100
rows of the counter table.

Drawing a pretty picture.