Nasty bug 🪳
First and foremost I fixed this ugly bug. Given this code:
```rust
    async fn connect(self, connection_string: string) -> Future<Result<Pool, Error>> {
        let pool = self
            .plugin
            .invoke_func_mut("sqlx", "Sqlx", "connect", [connection_string]);

        match pool {
            Result::Ok(pool) => Result::<Pool, Error>::Ok(Pool {
                plugin: self.plugin,
                inner: pool,
            }),
            Result::Err(e) => Result::<Pool, Error>::Err(Error {
                plugin: self.plugin,
                inner: e,
            }),
        }
    }
```

The error condition would cause the VM to bug out. It was really confusing
as to why. It turned out that `pool`, the `Result` was being overwritten by
the `pool` in `Ok(pool)`.

I also added some better error messages and some other junk.