A real melange
I changed ``sqlx::Row::get` to return a `Result`. In order for
this to work, I needed `Result::unwrap`. Well, to get that I
needed a `halt` expression. So I added `halt`, and then `Result::unwrap`.`
Somewhere in all this mess I ended up needing to hack the way that
`Enum::Tuple` implements `PartialEq`. I'm not really proud. The problem
is that sometimes the String value of the type matches and they have
different ValueTypes.

To make the blog page work correctly I needed `String::replace`.

I also had to hack where we are loading the /file directory relative to.
It should be the root of the project, but I moved the source into /src,
so now we look in ../file. This'll get sorted somehow. I really need
to just fix how the extruder finds files.