Oopsie
I let this one get out of hand.

I did some messing about with the sqlx plugin for the article.
I'm not sure why I touched the http plugin, but I did.

I got static hashmaps working in a very limited manner. The key must
be a String. I'm not in love with the syntax either, and I really
don't know that they belong. I was in the process of getting insert
and get working. There is a huge problem with inserting into an empty
hashmap. An empty list too if it's type isn't specified. It's the
UBER_HACK stuff. It doesn't typecheck. I can probably hardcode something
I suppose.

I also fixed a problem with the VM and async. I tried running the
bench.ore file and got panics from Await in the VM. The task was gone.
After much hairpulling I realized that the task was None because the
thing was being cloned. So I fixed Value::Task to wrap the AsyncTask
in a RefType.