Fixed two bugs
One was that the compiler was emitting STORE_LOCAL instructions
when initializing the iterator variable in a for loop. This was
causing all sorts of self referential stuff to happen when the
storage location is `i` and it's being assigned `[i]`.

The other bug was that the iterator was incorrectly set to have
the type of the future's value, rather than the future itself.