Feature/db articles (#38)
* Merged stashed changes into branch.

* Merge stashed changes into post develop merge.

* Got variable routing working?

Don't know what to call it. It's when you are listening to /blog/{slug},
and when you are called back you are given slug. That.

* Got custom 404 error message.

It's baked into the http plugin, which is not really that cool, but
it's a start.

* Added String::replace

Did something to the sql plugin as well.

* A real melange

I changed ``sqlx::Row::get` to return a `Result`. In order for
this to work, I needed `Result::unwrap`. Well, to get that I
needed a `halt` expression. So I added `halt`, and then `Result::unwrap`.`
Somewhere in all this mess I ended up needing to hack the way that
`Enum::Tuple` implements `PartialEq`. I'm not really proud. The problem
is that sometimes the String value of the type matches and they have
different ValueTypes.

To make the blog page work correctly I needed `String::replace`.

I also had to hack where we are loading the /file directory relative to.
It should be the root of the project, but I moved the source into /src,
so now we look in ../file. This'll get sorted somehow. I really need
to just fix how the extruder finds files.

* Fixed plugin xtask

It was failing when there was no misc directory.

* Got default html body and head from database

Had to fix a bug with if let expressions. The parser was not
generating a default for the else branch.

* Misc stuff

Just clean up during writing. New 404 background. Dunno.

* Change plugin to name in install script.

* Lot's of Plugin stuff

I started out fixing the Plugin loading section of the call expression
in the compiler. I was returning None for the type and noticed that
there was some commented out code. I think I was there researching my
article.

Anyway, I fixed the commented out code, by doing a fully qualified type
lookup in LuDog. So that fixed that and broke a bunch of plugins because
I hadn't named things correctly.

I also added mermaid.js support to the markdown plugin.

* Changed SQL testing

I changed the setup, and the table. It now has an INT, and I test
binding to it. Sadly it works.

* Changed SQL testing

I changed the setup, and the table. It now has an INT, and I test
binding to it. Sadly it works.

Missed a file.

* Cleanup -- comments.

* Internal Markdown Links

Got those working in the md plugin. Sweet.

Some cleanup and comments also as I write.

* Add uri query to http plugin

Also did a lot of commenting to the SQLX plugin.

* Oopsie

I let this one get out of hand.

I did some messing about with the sqlx plugin for the article.
I'm not sure why I touched the http plugin, but I did.

I got static hashmaps working in a very limited manner. The key must
be a String. I'm not in love with the syntax either, and I really
don't know that they belong. I was in the process of getting insert
and get working. There is a huge problem with inserting into an empty
hashmap. An empty list too if it's type isn't specified. It's the
UBER_HACK stuff. It doesn't typecheck. I can probably hardcode something
I suppose.

I also fixed a problem with the VM and async. I tried running the
bench.ore file and got panics from Await in the VM. The task was gone.
After much hairpulling I realized that the task was None because the
thing was being cloned. So I fixed Value::Task to wrap the AsyncTask
in a RefType.

* Fixed two bugs

One was that the compiler was emitting STORE_LOCAL instructions
when initializing the iterator variable in a for loop. This was
causing all sorts of self referential stuff to happen when the
storage location is `i` and it's being assigned `[i]`.

The other bug was that the iterator was incorrectly set to have
the type of the future's value, rather than the future itself.

* Last bug for bench.ore

This one was actually a bug in dwarf code. I called invoke_func
and needed to call invoke_func_mut.

* A Smattering of Things

Mostly I think it's getting the HashMap across the FFI boundary.

There is some commenting done in the sqlx plugin in the dwarf code.

* Added String method.