Really fixed that match issue
The fix I did broke other stuff. Specifically, a test where
a variable was reused in the test and it had different types.
The test failed because I was only grabbing the first type.

The real fix is to reverse the order of the values being pulled
from the blocks. But all of the values in the block need to be
maintained.

There is also some map related stuff in here. I think getting
map working was where this all started...